Report test coverage from script/test.sh
Because we're in docker, test coverage needs to be reported within the
same container; otherwise, we'll have to figure out how to share assets
across build steps. Because we're `set -e`ing this script, code coverage
shouldn't be instrumented unless the entire test suite passes, which the
typical behavior of code coverage tooling.