Update admin authorization to check team role vs Admin type
Whenever we check for admin authorization, check that the currently
authenticated user has the `team` role vs the `Admin` user type.

I chose to only change the underlying check logic and not the names of
functions or spec descriptions to keep the diff focused. Happy to
follow-up with those changes in a subsequent pull request.